+++ package.json
@@ -34,7 +34,7 @@
     "build": "pnpm -r run build",
     "lint": "oxlint --type-aware",
     "typecheck": "pnpm clean && pnpm -r run typecheck",
-    "pretest": "pnpm -r run pretest && pnpm run clean",
+    "pretest": "pnpm run clean && pnpm -r run pretest",
     "test": "vitest run --bail 1 --retry 2",
     "test:cov": "pnpm run test --coverage",
     "ci": "pnpm run test:cov",
