+++ pnpm-workspace.yaml
@@ -79,7 +79,6 @@ catalog:
   cluster-reload: ^2.0.0
   co-busboy: ^2.0.1
   coffee: '5'
-  common-bin: ^3.0.1
   content-disposition: ~0.5.4
   content-type: ^1.0.5
   cookie: ^1.0.2
@@ -94,7 +93,7 @@ catalog:
   dayjs: ^1.11.13
   debounce: ^2.2.0
   destroy: ^1.0.4
-  detect-port: ^2.1.0
+  detect-port: 2.1.0
   dumi: ^2.4.17
   egg-errors: ^2.3.2
   egg-logger: ^3.5.0
