--- tools/egg-bin/vitest.config.ts
@@ -1,7 +1,0 @@
-import { defineProject } from 'vitest/config';
-
-export default defineProject({
-  test: {
-    exclude: ['**/node_modules/**', '**/dist/**', '**/test/**'],
-  },
-});
