+++ tools/scripts/scripts/start-cluster.cjs
@@ -1,15 +1,22 @@
-/* eslint-disable @typescript-eslint/no-var-requires */
 const { debuglog } = require('node:util');
+
 const { importModule } = require('@eggjs/utils');
 
-const debug = debuglog('@eggjs/scripts/scripts/start-cluster');
+const debug = debuglog('egg/scripts/start-cluster/cjs');
 
 async function main() {
   debug('argv: %o', process.argv);
   const options = JSON.parse(process.argv[2]);
   debug('start cluster options: %o', options);
-  const { startCluster } = await importModule(options.framework);
+  const exports = await importModule(options.framework);
+  let startCluster = exports.startCluster;
+  if (typeof startCluster !== 'function') {
+    startCluster = exports.default.startCluster;
+  }
   await startCluster(options);
 }
 
-void main();
+main().catch(err => {
+  console.error(err);
+  process.exit(1);
+});
