+++ tools/scripts/src/index.ts
@@ -1,6 +1,6 @@
-import Start from './commands/start.js';
+import Start from './commands/start.ts';
 
 // exports.StopCommand = require('./lib/cmd/stop');
 
-export * from './baseCommand.js';
+export * from './baseCommand.ts';
 export { Start, Start as StartCommand };
