+++ tools/scripts/test/egg-scripts.test.ts
@@ -1,9 +1,10 @@
+import path from 'node:path';
+
 import { describe, it } from 'vitest';
 import coffee from 'coffee';
-import { getSourceFilename } from '../src/helper.ts';
 
 describe('test/egg-scripts.test.ts', () => {
-  const eggBin = getSourceFilename('../bin/run.js');
+  const eggBin = path.join(import.meta.dirname, '../bin/run.js');
 
   it('show help', async () => {
     await coffee
