+++ tools/scripts/test/fixtures/example-stop/app.js
@@ -1,0 +1,4 @@
+module.exports = () => {
+  // --no-deprecation
+  new Buffer('aaa');
+};
