+++ tools/scripts/test/fixtures/example-stop/config/config.default.js
@@ -1,0 +1,8 @@
+'use strict';
+
+exports.keys = '123456';
+
+exports.logger = {
+  level: 'WARN',
+  consoleLevel: 'WARN',
+};
