+++ tools/scripts/test/fixtures/example-stop/config/config.pre.js
@@ -1,0 +1,3 @@
+'use strict';
+
+exports.pre = true;
