+++ tools/scripts/test/ts.test.ts
@@ -1,21 +1,21 @@
 import path from 'node:path';
-import { fileURLToPath } from 'node:url';
-import { strict as assert } from 'node:assert';
 import fs from 'node:fs/promises';
 import cp from 'node:child_process';
 import { scheduler } from 'node:timers/promises';
-import { describe, it, beforeAll, afterAll, afterEach } from 'vitest';
+
+import { describe, it, beforeAll, afterAll, beforeEach, afterEach, expect } from 'vitest';
 import coffee from 'coffee';
 import { request } from 'urllib';
 import { mm, restore } from 'mm';
-import { cleanup, replaceWeakRefMessage, Coffee } from './utils.ts';
-import { isWindows, getSourceFilename } from '../src/helper.ts';
+import { detectPort } from 'detect-port';
 
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = path.dirname(__filename);
+import { cleanup, type Coffee } from './utils.ts';
+import { isWindows } from '../src/helper.ts';
+
+const __dirname = import.meta.dirname;
 
 describe('test/ts.test.ts', () => {
-  const eggBin = getSourceFilename('../bin/run.js');
+  const eggBin = path.join(__dirname, '../bin/run.js');
   const homePath = path.join(__dirname, 'fixtures/home');
   const waitTime = 5000;
   let fixturePath: string;
@@ -35,7 +35,7 @@ describe('test/ts.test.ts', () => {
         cwd: fixturePath,
         shell: isWindows,
       });
-      assert.equal(result.stderr.toString(), '');
+      expect(result.stderr.toString()).toBe('');
     });
 
     afterEach(async () => {
@@ -44,45 +44,48 @@ describe('test/ts.test.ts', () => {
     });
 
     it('--ts', async () => {
-      app = coffee.fork(eggBin, ['start', '--workers=1', '--ts', fixturePath]) as Coffee;
-      // app.debug();
+      const port = await detectPort();
+      app = coffee.fork(eggBin, ['start', '--workers=1', '--ts', `--port=${port}`, fixturePath]) as Coffee;
+      app.debug();
       app.expect('code', 0);
 
       await scheduler.wait(waitTime);
 
-      assert.equal(replaceWeakRefMessage(app.stderr), '');
-      assert.match(app.stdout, /egg started on http:\/\/127\.0\.0\.1:7001/);
-      const result = await request('http://127.0.0.1:7001', { dataType: 'json' });
+      // expect(replaceWeakRefMessage(app.stderr)).toBe('');
+      expect(app.stdout).toMatch(/egg started on http:\/\/127\.0\.0\.1:\d+/);
+      const result = await request(`http://127.0.0.1:${port}`, { dataType: 'json' });
       // console.log(result.data);
-      assert(result.data.stack.includes(path.normalize('app/controller/home.ts:6:13')));
+      expect(result.data.stack).toContain(path.normalize('app/controller/home.ts:6:13'));
     });
 
     it('--typescript', async () => {
-      app = coffee.fork(eggBin, ['start', '--workers=1', '--typescript', fixturePath]) as Coffee;
+      const port = await detectPort();
+      app = coffee.fork(eggBin, ['start', '--workers=1', '--typescript', `--port=${port}`, fixturePath]) as Coffee;
       // app.debug();
       app.expect('code', 0);
 
       await scheduler.wait(waitTime);
 
-      assert.equal(replaceWeakRefMessage(app.stderr), '');
-      assert.match(app.stdout, /egg started on http:\/\/127\.0\.0\.1:7001/);
-      const result = await request('http://127.0.0.1:7001', { dataType: 'json' });
+      // expect(replaceWeakRefMessage(app.stderr)).toBe('');
+      expect(app.stdout).toMatch(/egg started on http:\/\/127\.0\.0\.1:\d+/);
+      const result = await request(`http://127.0.0.1:${port}`, { dataType: 'json' });
       // console.log(result.data);
-      assert(result.data.stack.includes(path.normalize('app/controller/home.ts:6:13')));
+      expect(result.data.stack).toContain(path.normalize('app/controller/home.ts:6:13'));
     });
 
     it('--sourcemap', async () => {
-      app = coffee.fork(eggBin, ['start', '--workers=1', '--sourcemap', fixturePath]) as Coffee;
+      const port = await detectPort();
+      app = coffee.fork(eggBin, ['start', '--workers=1', '--sourcemap', `--port=${port}`, fixturePath]) as Coffee;
       // app.debug();
       app.expect('code', 0);
 
       await scheduler.wait(waitTime);
 
-      assert.equal(replaceWeakRefMessage(app.stderr), '');
-      assert.match(app.stdout, /egg started on http:\/\/127\.0\.0\.1:7001/);
-      const result = await request('http://127.0.0.1:7001', { dataType: 'json' });
+      // expect(replaceWeakRefMessage(app.stderr)).toBe('');
+      expect(app.stdout).toMatch(/egg started on http:\/\/127\.0\.0\.1:\d+/);
+      const result = await request(`http://127.0.0.1:${port}`, { dataType: 'json' });
       // console.log(result.data);
-      assert(result.data.stack.includes(path.normalize('app/controller/home.ts:6:13')));
+      expect(result.data.stack).toContain(path.normalize('app/controller/home.ts:6:13'));
     });
   });
 
@@ -95,7 +98,7 @@ describe('test/ts.test.ts', () => {
         cwd: fixturePath,
         shell: isWindows,
       });
-      assert.equal(result.stderr.toString(), '');
+      expect(result.stderr.toString()).toBe('');
     });
 
     afterEach(async () => {
@@ -104,17 +107,18 @@ describe('test/ts.test.ts', () => {
     });
 
     it('should got correct stack', async () => {
-      app = coffee.fork(eggBin, ['start', '--workers=1', fixturePath]) as Coffee;
+      const port = await detectPort();
+      app = coffee.fork(eggBin, ['start', '--workers=1', `--port=${port}`, fixturePath]) as Coffee;
       // app.debug();
       app.expect('code', 0);
 
       await scheduler.wait(waitTime);
 
-      assert.equal(replaceWeakRefMessage(app.stderr), '');
-      assert.match(app.stdout, /egg started on http:\/\/127\.0\.0\.1:7001/);
-      const result = await request('http://127.0.0.1:7001', { dataType: 'json' });
-      console.log(result.data);
-      assert.match(result.data.stack, /home\.ts:6:13/);
+      // expect(replaceWeakRefMessage(app.stderr)).toBe('');
+      expect(app.stdout).toMatch(/egg started on http:\/\/127\.0\.0\.1:\d+/);
+      const result = await request(`http://127.0.0.1:${port}`, { dataType: 'json' });
+      // console.log(result.data);
+      expect(result.data.stack).toMatch(/home\.ts:6:13/);
       // assert(result.data.stack.includes(path.normalize('app/controller/home.ts:6:13')));
     });
   });
