+++ tools/scripts/tsdown.config.ts
@@ -2,7 +2,7 @@ import { defineConfig } from 'tsdown';
 
 export default defineConfig({
   entry: 'src/**/*.ts',
-  // unbundle: true,
+  unbundle: true,
   unused: {
     level: 'error',
   },
