+++ vitest.config.ts
@@ -5,7 +5,7 @@ export default defineConfig({
     projects: [
       'packages/*',
       'plugins/*',
-      'tools/*',
+      'tools/create-egg',
       // FIXME: enable this will cause one test file run twice
       // {
       //   extends: true,
